refactor(application): split lsp runtime by projection seams
`lsp_runtime.rs` (4.4k lines) carried the registered project authority,
diagnostic and test-run projection, the concrete feedback source, the
context projection helpers, document path admission, and four inline
test modules. Move each, unchanged, into `lsp_runtime/`:

- `registered_authority`: `RegisteredProjectLspAuthority` and the
  projection scope port
- `diagnostic_projection`: feedback diagnostic admission and the
  diagnostics-store projection
- `managed_test_runs`: the operation-event test-run projection
- `feedback_source`: `ConcreteFeedbackLspSource` and its context
  expansion records
- `context_projection`: projection composition helpers
- `document_paths`: path validation and project file access
- `path_tests`, `context_expansion_tests`, `projection_tests`,
  `diagnostic_admission_tests`: the former inline test modules

The root keeps the scope/queue/expansion record types, the session
factory, and re-exports every public name at its old path. Moved
root-private items become `pub(super)`, the scope they already had.
